Refactor: Reorganize classes into new packages for better code readability
- Moved `Sorter` and `InsertionSorter` classes to the new package `ua.nure.cpp.lab1.computations.sorters`.
- Moved `ArrayReader`, `ConsoleArrayReader`, and `InputStreamArrayReader` classes to the new package `ua.nure.cpp.lab1.computations.readers`.
