+++ lab1/src/main/java/ua/nure/cpp/lab1/computations/readers/ArrayReader.java
@@ -1,4 +1,4 @@
-package ua.nure.cpp.lab1.computations;
+package ua.nure.cpp.lab1.computations.readers;
 
 public interface ArrayReader {
     int[] read();
